PIm-7211: Convert section to typescript
