+++ src/Pim/Bundle/EnrichBundle/Resources/config/requirejs.yml
@@ -641,7 +641,7 @@ config:
         pim/group-type-edit-form/delete:                 pimenrich/js/group-type/form/delete
 
         # Catalog volume monitoring
-        pim/catalog-volume/section:                      pimenrich/js/catalog-volume/section
+        pim/catalog-volume/section:                      pimenrich/js/catalog-volume/section.ts
         pim/catalog-volume/header:                       pimenrich/js/catalog-volume/header.ts
 
         # Channel
